Tighten metric sandboxing and form validation
- Lift the .constructor-chain block (`runSandboxed`) and shadowed-globals
  list to a shared `simulation/sandbox.ts` and apply them to compileMetric
  so it matches compileScenario's hardening. Adds a regression test that
  `({}).constructor.constructor("…")()` is rejected.
- Reject empty metric code via a form-level validator so the Save button
  surfaces "Metric code is required" instead of letting the form save a
  metric that compileMetric would later reject at runtime.
- Gate ViewMetricDrawer's `open` on the selected metric still existing —
  switching the timeline picker while the edit drawer is open no longer
  leaves an empty overlay behind.
